fix: move model tests to its own packages
